# openapiBorsalino.yaml
...
openapi: "3.0.1"
info:
  title: "Delilah Resto"
  description: "Borsalino backend API."
  contact:
    email: "federicoborsalino@gmail.com"
  version: "1.0.0"
servers:
- url: "https://virtserver.swaggerhub.com/DefenderAndAttackHack/delilaResto/1.0.0"
  description: "SwaggerHub API Auto Mocking"
- url: "https://localhost:3000/v1"
- url: "http://localhost:3000/v1"
security:
- bearerAuth: []
...
paths:
  /users:
    get:
      tags:
      - "Usuario"
      summary: "obtener lista de usuarios(Admin) o detalles de la cuenta de usuario(not\
        \ admin)"
      description: "lista de usuario"
      operationId: "getUserInfo"
      responses:
        "200":
          description: "Successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Send"
              examples:
                Token Invalidation:
                  summary: "Unauthorized - Invalid Token"
                  value: "Unauthorized - Invalid Token"
                Disabled Account:
                  summary: "Unauthorized - Account is Disabled"
                  value: "Unauthorized - User account is disabled"
        "404":
          description: "The specified resource was not found"
        default:
          description: "Unexpected error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags:
      - "Usuario"
      summary: "editar datos de usuario activos"
      operationId: "editActiveUserData"
      requestBody:
        description: "Fields to change"
        content:
          application/json:
            schema:
              allOf:
              - $ref: "#/components/schemas/User"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "nombre de usuario o contraseña not valida"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Send"
              examples:
                Token Invalidation:
                  summary: "Unauthorized - Invalid Token"
                  value: "Unauthorized - Invalid Token"
                Disabled Account:
                  summary: "Unauthorized - Account is Disabled"
                  value: "Unauthorized - User account is disabled"
        "404":
          description: "The specified resource was not found"
        "409":
          description: "el usuario o email ya existen"
        default:
          description: "Unexpected error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
      - "Usuario"
      summary: "crear usuario"
      description: "crear usuario en base de datos"
      operationId: "crear usuario"
      requestBody:
        description: "crear objetos de usuario"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "200":
          description: "usuario agregado correctamente a la base de datos"
        "400":
          description: "error al validar los datos de entrada"
        "409":
          description: "usuario existente elegir otro nombre"
        default:
          description: "Unexpected error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: []
    delete:
      tags:
      - "Usuario"
      summary: "Cuenta de usuario deshabilitada (validacion via token)"
      operationId: "deleteActiveUser"
      responses:
        "200":
          description: "User account disabled"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Send"
              examples:
                Token Invalidation:
                  summary: "Unauthorized - Invalid Token"
                  value: "Unauthorized - Invalid Token"
                Disabled Account:
                  summary: "Unauthorized - Account is Disabled"
                  value: "Unauthorized - User account is disabled"
        default:
          description: "Unexpected error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/login:
    get:
      tags:
      - "Usuario"
      summary: "registrar al usuario en el sistema"
      operationId: "loginUser"
      parameters:
      - name: "nombre de usuario o email"
        in: "query"
        description: "nombre de usuario o email para login"
        required: true
        style: "form"
        explode: true
        schema:
          type: "string"
      - name: "password"
        in: "query"
        description: "password para login"
        required: true
        style: "form"
        explode: true
        schema:
          type: "string"
      responses:
        "200":
          description: "operacion exitosa"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        "400":
          description: "nombre de usuario o contraseña not valida"
        "401":
          description: "Invalid request, user account is disabled"
        default:
          description: "Unexpected error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: []
  /users/:username:
    get:
      tags:
      - "Usuario"
      summary: "obtener usuario"
      operationId: "getUserByName"
      parameters:
      - name: "username"
        in: "path"
        description: "nombre para buscar"
        required: true
        style: "simple"
        explode: false
        schema:
          type: "string"
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Send"
              examples:
                Token Invalidation:
                  summary: "Unauthorized - Invalid Token"
                  value: "Unauthorized - Invalid Token"
                Disabled Account:
                  summary: "Unauthorized - Account is Disabled"
                  value: "Unauthorized - User account is disabled"
                Not Admin:
                  summary: "Unauthorized - Not an admin"
                  value: "Unauthorized - Not an admin"
        "404":
          description: "The specified resource was not found"
        default:
          description: "Unexpected error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags:
      - "Usuario"
      summary: "editar por nombre de usuario (solo admin)"
      description: "Edita un nombre de usuario, obtenido por nombre de usuario (solo\
        \ administrador). Se puede enviar el objeto completo o solo los parámetros\
        \ que se deben modificar."
      operationId: "updateUserAdmin"
      parameters:
      - name: "username"
        in: "path"
        description: "nombre que necesita ser actualizado"
        required: true
        style: "simple"
        explode: false
        schema:
          type: "string"
      requestBody:
        description: "objeto de usuario actualizado"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: "User {username} was modified correctly"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Send"
              examples:
                Token Invalidation:
                  summary: "Unauthorized - Invalid Token"
                  value: "Unauthorized - Invalid Token"
                Disabled Account:
                  summary: "Unauthorized - Account is Disabled"
                  value: "Unauthorized - User account is disabled"
                Not Admin:
                  summary: "Unauthorized - Not an admin"
                  value: "Unauthorized - Not an admin"
        "404":
          description: "The specified resource was not found"
        default:
          description: "Unexpected error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags:
      - "Usuario"
      summary: "eliminar usuario (solo admin)"
      description: "Esto solo puede hacerlo un administrador que haya iniciado sesió\
        n."
      operationId: "deleteUserAdmin"
      parameters:
      - name: "username"
        in: "path"
        description: "El nombre que debe eliminarse"
        required: true
        style: "simple"
        explode: false
        schema:
          type: "string"
      responses:
        "200":
          description: "User {username} was disabled correctly"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Send"
              examples:
                Token Invalidation:
                  summary: "Unauthorized - Invalid Token"
                  value: "Unauthorized - Invalid Token"
                Disabled Account:
                  summary: "Unauthorized - Account is Disabled"
                  value: "Unauthorized - User account is disabled"
                Not Admin:
                  summary: "Unauthorized - Not an admin"
                  value: "Unauthorized - Not an admin"
        "404":
          description: "The specified resource was not found"
        default:
          description: "Unexpected error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    User:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 100
        username:
          type: "string"
          example: "Fede Borsalino"
        password:
          type: "string"
          example: "password1"
        email:
          type: "string"
          example: "fedeBorsa@gmail.com"
        delivery_address:
          type: "string"
          example: "duarte quiros 3100"
        full_name:
          type: "string"
          example: "Federico Borsalino"
        phone:
          type: "integer"
          example: 543519999999
    UserInfo:
      type: "array"
      items:
        type: "object"
        properties:
          user_id:
            type: "integer"
            example: 1
          username:
            type: "string"
            example: "fborsalino"
          full_name:
            type: "string"
            example: "Fernando Borsalino"
          email:
            type: "string"
            example: "fborsalino@gmail.com"
          phone:
            type: "integer"
            example: 543513999999
          delivery_address:
            type: "string"
            example: "colon 5300"
          is_admin:
            type: "boolean"
            example: false
          is_disabled:
            type: "boolean"
            example: false
...
    Token:
      type: "object"
      properties:
        token:
          type: "string"
          example: "TOKEN999320"
    Error:
      required:
      - "error"
      type: "object"
      properties:
        error:
          type: "string"
          example: "ocurrio un error inesperado"
    Send:
      type: "string"
      example: "Error message"
...
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"